Add third party data to network pages
