+++ pages/api/cloudflare.js
@@ -4,7 +4,14 @@ import dayjs from 'services/dayjs'
 export default function handler(req, res) {
   const dateStart = req.query?.from
   const dateEnd = req.query?.to
-  const location = req.query?.country
+  const location = req.query?.location
+
+  if (!dayjs(dateStart).isValid() || dayjs(dateStart).isAfter(dayjs())) {
+    return res.status(400).json('Invalid Start Date')
+  }
+  if (!dayjs(dateEnd).isValid() || dayjs(dateEnd).isAfter(dayjs())) {
+    return res.status(400).json('Invalid End Date')
+  }
 
   const diff = dayjs(dateEnd).diff(dayjs(dateStart), 'day')
   const aggInterval = diff <= 30 ? '1h' : '1d'
@@ -12,20 +19,6 @@ export default function handler(req, res) {
   axios({
     method:'get',
     url:`https://api.cloudflare.com/client/v4/radar/netflows/timeseries?name=all&product=all&dateStart=${dateStart}&dateEnd=${dateEnd}&location=${location}&name=http&product=http&dateStart=${dateStart}&dateEnd=${dateEnd}&location=${location}&aggInterval=${aggInterval}&normalization=MIN0_MAX`,
-
-    
-    
-    // params: {
-    //   aggInterval: '1h',
-    //   dateRange: '7d',
-    //   location: 'RU',
-    //   // product: ['http', 'all'],
-    //   name: 'bla'
-    // },
-    // paramsSerializer: params => {
-    //   return qs.stringify(params)
-    // },
-    // baseURL: 'https://api.cloudflare.com/',
     headers: {
       'X-Auth-Key': process.env.CLOUDFLARE_TOKEN,
       'X-Auth-Email': process.env.CLOUDFLARE_EMAIL
@@ -35,6 +28,6 @@ export default function handler(req, res) {
     return res.status(200).json(data)
   }).catch((err) =>{
     console.log('error', err.response.data)
-    return res.status(200).json(err)
+    return res.status(400).json(err)
   }) 
 }
