+++ pages/country/[countryCode].js
@@ -12,7 +12,6 @@ import { useIntl } from 'react-intl'
 import { StickyContainer, Sticky } from 'react-sticky'
 import { getLocalisedRegionName } from '../../utils/i18nCountries'
 import dayjs from 'services/dayjs'
-import { ResponsiveLine } from '@nivo/line'
 
 import Form from 'components/network/Form'
 import NavBar from 'components/NavBar'
@@ -23,6 +22,7 @@ import WebsitesSection from 'components/country/Websites'
 import AppsSection from 'components/country/Apps'
 import { CountryContextProvider } from 'components/country/CountryContext'
 import CountryHead from 'components/country/CountryHead'
+import ThirdPartyDataGraph from 'components/ThirdPartyDataGraph'
 
 const getCountryReports = (countryCode, data) => {
   const reports = data.filter((article) => (
@@ -143,58 +143,6 @@ const Country = ({ countryCode, overviewStats, reports, ...coverageDataSSR }) =>
 
   const { testCoverage, networkCoverage } = newData !== false ? newData : coverageDataSSR
 
-  const [graphData, setGraphData] = useState([])
-
-  const cloudflareData = useEffect(() => {
-    if (query.until && query.until) {
-      const to = dayjs.utc(query.until)
-      const from = dayjs.utc(query.since)
-
-      axios({
-        method: 'get',
-        url: `/api/cloudflare?from=${from.toISOString().split('.')[0]+'Z'}&to=${to.toISOString().split('.')[0]+'Z'}&country=${countryCode}`,
-      }).then(({data}) => {
-        const ruData = data.result.all.timestamps.map((st, i) => {
-          return {
-            'x': st,
-            'y': Number(data.result.all.values[i])
-          }
-        })
-        setGraphData((oldVal) => {
-          return [...oldVal, {
-            'id': 'Cloudflare',
-            // 'color': 'hsl(0, 100%, 50%)',
-            'data': ruData,
-          }]
-        })
-      }).catch(() => {})
-  
-      axios({
-          method: 'get',
-          url: `https://api.ioda.inetintel.cc.gatech.edu/v2/signals/raw/country/${countryCode}?from=${Math.round(from.valueOf()/1000)}&until=${Math.round(to.valueOf()/1000)}&sourceParams=WEB_SEARCH`,
-      }).then(({data}) => {
-        const graphData2 = data.data[0].map((item) => {
-          const max = Math.max(...item.values)
-          const values = item.values.map((val, i) => {
-            const time = dayjs(item.from * 1000).add(item.step * i, 'second').utc().toISOString().split('.')[0]+'Z'
-            return {x: time, y: val ? val / max : null}
-          })
-          return {
-            'id': item.datasource,
-            // 'color': 'hsl(100, 70%, 50%)',
-            'data': values
-          }
-        })
-  
-        setGraphData((oldVal) => {
-          return [...oldVal, ...graphData2]
-        })
-      })
-    }
-  }, [])
-
-  console.log('graphData', graphData)
-
   return (
     <>
       <CountryHead countryName={countryName} measurementCount={overviewStats.measurement_count} measuredSince={overviewStats.first_bucket_date} networkCount={overviewStats.network_count} />
@@ -248,108 +196,14 @@ const Country = ({ countryCode, overviewStats, reports, ...coverageDataSSR }) =>
                 <AppsSection />
               </CountryContextProvider>
             </Box>
-            <Box style={{width: '100%',height: '500px'}}>
-              {!!graphData.length && 
-                  <ResponsiveLine
-                  data={graphData}
-                  margin={{ top: 50, right: 20, bottom: 70, left: 30 }}
-                  enablePoints={false}
-                  lineWidth={1}
-                  xScale={{
-                    type: 'time',
-                    format: '%Y-%m-%dT%H:%M:%SZ',
-                    precision: 'minute',
-                    useUTC: true,
-                  }}
-                  yScale={{
-                    type: 'linear',
-                    stacked: false,
-                    min: 0,
-                    max: 1
-                  }}
-                  // xFormat="time:%Hh"
-                  axisBottom={{
-                    format: '%Y-%m-%d',
-                  }}
-                  // enableSlices='x'
-                  useMesh={true}
-
-
-
-                  // data={graphData}
-                  // curve="monotoneX"
-                  // enableSlices='x'
-                  // animate={true}
-                  // margin={{ top: 50, right: 110, bottom: 50, left: 60 }}
-                  // xScale={{ 
-                  //   type: 'time',
-                  //   format: '%Y-%m-%d',
-                  //   useUTC: true,
-                  //   precision: 'day',
-                  // }}
-                  // yScale={{
-                  //     type: 'linear',
-                  //     min: 'auto',
-                  //     max: 'auto',
-                  //     stacked: false,
-                  //     reverse: false
-                  // }}
-                  // // yFormat=" >-.2f"
-                  // // axisTop={null}
-                  // // axisRight={null}
-                  // axisBottom={{
-                  //   format: '%Y-%m-%d',
-                  //   orient: 'bottom',
-                  //   tickSize: 5,
-                  //   tickPadding: 5,
-                  //   tickRotation: 0,
-                  //   legend: 'date',
-                  //   legendOffset: 36,
-                  //   legendPosition: 'middle',
-                  //   tickValues: 6
-                  // }}
-                  // axisLeft={{
-                  //     orient: 'left',
-                  //     tickSize: 5,
-                  //     tickPadding: 5,
-                  //     tickRotation: 0,
-                  //     legend: 'count',
-                  //     legendOffset: -40,
-                  //     legendPosition: 'middle'
-                  // }}
-                  // pointSize={10}
-                  // pointColor={{ theme: 'background' }}
-                  // pointBorderWidth={2}
-                  // pointBorderColor={{ from: 'serieColor' }}
-                  // pointLabelYOffset={-12}
-              
-                  legends={[
-                    {
-                      anchor: 'bottom-left',
-                      direction: 'row',
-                      // justify: false,
-                      // translateX: 100,
-                      translateY: 70,
-                      itemsSpacing: 0,
-                      itemDirection: 'left-to-right',
-                      itemWidth: 90,
-                      itemHeight: 20,
-                      // itemOpacity: 0.75,
-                      symbolSize: 12,
-                      symbolShape: 'circle',
-                      symbolBorderColor: 'rgba(0, 0, 0, .5)',
-                      // effects: [{
-                      //   on: 'hover',
-                      //   style: {
-                      //     itemBackground: 'rgba(0, 0, 0, .03)',
-                      //     itemOpacity: 1
-                      //   }
-                      // }]
-                    }
-                  ]}
+            
+              {query.since && query.until && 
+                <ThirdPartyDataGraph
+                  country={countryCode}
+                  since={query.since}
+                  until={query.until}
                 />
               }
-            </Box>
           </Flex>
         </Container>
       </StickyContainer>
